+++ src/Common/src/uk/ac/ox/zoo/seeg/abraid/mp/common/domain/Location.java
@@ -60,7 +60,7 @@ public class Location {
 
     // The HealthMap country ID (if any).
     @Column(name = "healthmap_country_id")
-    private Long healthMapCountryId;
+    private Integer healthMapCountryId;
 
     // The GAUL code of the admin 1/2 unit assigned to the location during QC stage 1.
     @Column(name = "admin_unit_qc_gaul_code")
@@ -103,7 +103,7 @@ public class Location {
         this.name = name;
     }
 
-    public Location(String name, double x, double y, LocationPrecision precision, long healthMapCountryId) {
+    public Location(String name, double x, double y, LocationPrecision precision, int healthMapCountryId) {
         this(name, x, y, precision);
         this.healthMapCountryId = healthMapCountryId;
     }
@@ -144,11 +144,11 @@ public class Location {
         this.resolutionWeighting = resolutionWeighting;
     }
 
-    public Long getHealthMapCountryId() {
+    public Integer getHealthMapCountryId() {
         return healthMapCountryId;
     }
 
-    public void setHealthMapCountryId(Long healthMapCountryId) {
+    public void setHealthMapCountryId(Integer healthMapCountryId) {
         this.healthMapCountryId = healthMapCountryId;
     }
